% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{ADPAlphaCron}
\alias{ADPAlphaCron}
\title{Alfa de Cronbach}
\usage{
ADPAlphaCron(X)
}
\arguments{
\item{X}{Data frame ou matriz numérica com variáveis contínuas.}
}
\value{
Uma lista contendo:
\itemize{
\item \code{alpha_all}: objeto completo retornado pela função \code{psych::alpha}.
\item \code{alpha_df}: data frame com os valores de alfa caso cada variável seja excluída.
\item \code{alpha_drop}: vetor com os valores de alfa após remoção de cada variável.
\item \code{alpha_total}: valor total do alfa de Cronbach para o conjunto.
}
}
\description{
Calcula o coeficiente alfa de Cronbach para medir a consistência interna
de um conjunto de variáveis, além do impacto da remoção de cada variável.
}
\examples{
ADPAlphaCron(mtcars)

}
