% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPBoxCox.R
\name{ADPBoxCox}
\alias{ADPBoxCox}
\title{Aplicar transformacao Box-Cox condicionalmente}
\usage{
ADPBoxCox(
  dadoswin = NULL,
  dados_originais = NULL,
  classe = NULL,
  nome = NULL,
  metodo = NULL
)
}
\arguments{
\item{dadoswin}{\code{data.frame} com os dados ja winsorizados.}

\item{dados_originais}{\code{data.frame} com os dados originais.}

\item{classe}{Vetor de caracteres indicando a classe de cada variavel.}

\item{nome}{Vetor com os nomes das variaveis.}

\item{metodo}{String indicando o metodo de transformacao:
\itemize{
\item \code{"forecast"} - usa \code{forecast::BoxCox}.
\item \code{"COINr"} - usa \code{COINr::boxcox}.
\item \code{"yeojohnson"} - usa \code{bestNormalize::yeojohnson}.
}}
}
\value{
Lista com:
\describe{
\item{meta}{\code{data.frame} com informacoes sobre a decisao de aplicar Box-Cox para cada variavel.}
\item{data}{\code{data.frame} com os dados transformados (ou originais, caso nao se aplique Box-Cox).}
}
}
\description{
Aplica a transformacao Box-Cox (ou variacoes) a variaveis numericas de acordo com criterios de distorcao (skewness) e curtose.
Apenas variaveis com distorcao >= 2 e curtose >= 3.5 sao transformadas.
}
\details{
A decisao de aplicar Box-Cox e baseada nas seguintes regras:
\itemize{
\item Se \code{skewness} (distorcao) >= 2 \strong{e}
\item \code{kurtosis} (curtose) >= 3.5
entao a transformacao e aplicada.
}
}
\examples{
# Exemplo ficticio (assumindo que boxcox_transform ja exista)
set.seed(123)
dados <- data.frame(
  x1 = c(rnorm(10), 100),# variavel distorcida
  x2 = rnorm(11))        # variavel normal

meta_test <- data.frame(N = c(1,2),
                        NIVEL = c("7","7"),
                        CODE = c("x1","x2"),
                        NOME = c("Variavel exemplo x1",
                                  "Variavel exemplo x2"),
                        TIPO = c("Indicator","Indicator"),
                        PAI  = c("X1L2","X2L2"),
                        CLASSE =c("Numerico","Numerico"))

dados_win <- winsorize_apply(dataset=dados,
                              metadados=meta_test)

ADPBoxCox(dadoswin = dados_win$dataset, 
          dados_originais = dados, 
          classe = meta_test$CLASSE,
          nome = meta_test$CODE,
          metodo = "forecast")

}
