% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{ADPvif}
\alias{ADPvif}
\title{Fator de Inflação da Variância (VIF)}
\usage{
ADPvif(X)
}
\arguments{
\item{X}{Data frame ou matriz numérica com variáveis contínuas.}
}
\value{
Data frame com duas colunas:
\itemize{
\item \code{Indicador}: nome da variável.
\item \code{VIF}: valor do VIF arredondado para 3 casas decimais.
}
}
\description{
Calcula os valores de VIF para avaliar a multicolinearidade entre variáveis
de um conjunto de dados.
}
\examples{
ADPvif(mtcars)

}
