% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_files_office.r
\name{carregar_pptx}
\alias{carregar_pptx}
\title{Abrir modelo de apresentação do pacote}
\usage{
carregar_pptx()
}
\value{
Objeto \code{pptx} (classe do pacote officer).
}
\description{
Esta função carrega o arquivo .pptx incluído no pacote
e retorna um objeto pptx pronto para ser manipulado com o pacote officer.
}
\examples{
ppt <- carregar_pptx()
print(ppt)

}
