% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_files_office.r
\name{carregar_xlsx}
\alias{carregar_xlsx}
\title{Abrir dados de exemplo do pacote}
\usage{
carregar_xlsx()
}
\value{
Objeto \code{xlsx} (classe do pacote officer).
}
\description{
Esta função carrega o arquivo .xlsx incluído no pacote
e retorna um objeto xlsx pronto para ser manipulado com o pacote officer.
}
\examples{
xlsx_file <- carregar_xlsx()
print(xlsx_file)

}
