% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_flextable_normal.R
\name{cria_flextable_normalidade}
\alias{cria_flextable_normalidade}
\title{Cria um flextable boxcox a partir de uma tabela}
\usage{
cria_flextable_normalidade(tab_input, alpha = 0.05)
}
\arguments{
\item{tab_input}{Data frame a ser convertido para flextable.}

\item{alpha}{nivel de confianca}
}
\value{
Lista contendo:
\describe{
\item{ft}{flextable gerado}
\item{largura_colunas}{vetor com larguras utilizadas}
}
}
\description{
Calcula larguras das colunas automaticamente e retorna o flextable
configurado.
}
\examples{
\dontrun{
icode = "MMPD"
cria_flextable_normalidade(result$Data_Bxc$data[[icode]])
}

}
