% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_flextable_winsorize.R
\name{cria_flextable_winsorize}
\alias{cria_flextable_winsorize}
\title{Cria um flextable winsorize a partir de uma tabela}
\usage{
cria_flextable_winsorize(tab_input)
}
\arguments{
\item{tab_input}{Data frame a ser convertido para flextable.}
}
\value{
Lista contendo:
\describe{
\item{ft}{flextable gerado}
\item{largura_colunas}{vetor com larguras utilizadas}
}
}
\description{
Calcula larguras das colunas automaticamente e retorna o flextable
configurado.
}
\examples{
\dontrun{
cria_flextable_winsorize(mtcars)
}

}
