% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_ref}
\alias{data_ref}
\title{Metadados data_ref}
\format{
Um tibble com X linhas e Y colunas.
}
\source{
Fonte dos dados Base_inicial_SA_Acesso.xlsx
}
\usage{
data_ref
}
\description{
Colunas de referencia para indexacao do dataset.
}
\keyword{datasets}
