% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{datasetN7}
\alias{datasetN7}
\title{Dados de exemplo datasetN7}
\format{
Um tibble com X linhas e Y colunas.
}
\source{
Fonte dos dados Base_inicial_SA_Acesso.xlsx
}
\usage{
datasetN7
}
\description{
Um tibble de exemplo com dados do N7.
}
\keyword{datasets}
