% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{FigContNA}
\alias{FigContNA}
\title{Figura da Contagem de Valores Ausentes (NA)}
\usage{
FigContNA(Y, file = FALSE, nfile = "FigContNA.png", visivel = FALSE)
}
\arguments{
\item{Y}{Vetor nomeado com contagem de NAs por variável.}

\item{file}{NULL.}

\item{nfile}{Nome do arquivo de saída (PNG).}

\item{visivel}{logical, se TRUE o grafico e exibido no prompt.}
}
\value{
Nenhum objeto retornado. Salva a figura em arquivo PNG.
}
\description{
Gera grafico de barras mostrando a quantidade de valores ausentes por variável.
}
\examples{
cont_na <- apply(mtcars, 2, function(x) sum(is.na(x)))
FigContNA(cont_na, "ContNA.png",visivel=TRUE)
}
