% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{iData_N7}
\alias{iData_N7}
\title{Dados de exemplo iData_N7}
\format{
Um tibble com X linhas e Y colunas.
}
\source{
Fonte dos dados (se houver)
}
\usage{
iData_N7
}
\description{
Um tibble de exemplo com dados do N7.
}
\keyword{datasets}
