% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{iMeta_N7}
\alias{iMeta_N7}
\title{Metadados iMeta_N7}
\format{
Um tibble com X linhas e Y colunas.
}
\source{
Fonte dos dados (se houver)
}
\usage{
iMeta_N7
}
\description{
Metadados do dataset iData_N7.
}
\keyword{datasets}
