% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_result_normal.R
\name{map_result_normal}
\alias{map_result_normal}
\title{Mapa tematico para indicadores municipais}
\usage{
map_result_normal(
  iNome,
  DADOS,
  REFERENCE,
  titulo = "Titulo",
  shp_mun = NULL,
  shp_uf = NULL,
  salvar = FALSE,
  dir = NULL,
  nome_arquivo = NULL,
  largura = 20,
  altura = 24,
  dpi = 300,
  fsize = 16
)
}
\arguments{
\item{iNome}{Nome da coluna de DADOS a ser mapeada.}

\item{DADOS}{Data frame com valores.}

\item{REFERENCE}{Data frame com GEOCOD.}

\item{titulo}{Título do mapa.}

\item{shp_mun}{Shapefile municipal (sf). Se NULL, usa mun_shp interno.}

\item{shp_uf}{Shapefile estadual (sf). Se NULL, usa uf_shp interno.}

\item{salvar}{Se TRUE, salva o mapa.}

\item{dir}{diretorio em que o arquivo sera salvo}

\item{nome_arquivo}{Nome do arquivo salvo.}

\item{largura}{Largura do grafico salvo.}

\item{altura}{Altura do grafico salvo.}

\item{dpi}{Resolução.}

\item{fsize}{Tamanho da fonte.}
}
\value{
Objeto ggplot2.
}
\description{
Gera um mapa temático usando shapefiles internos do pacote ou fornecidos pelo usuario.
}
\examples{
\dontrun{
# Exemplo que usa os shapefiles internos
# map_result_normal("MMPD", dados_N7, ref_N7)
}
}
