% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{metadadosN7}
\alias{metadadosN7}
\title{Metadados metadadosN7}
\format{
Um tibble com X linhas e Y colunas.
}
\source{
Fonte dos dados Base_inicial_SA_Acesso.xlsx
}
\usage{
metadadosN7
}
\description{
Metadados do dataset datasetN7.
}
\keyword{datasets}
