% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo_slides.r
\name{modelo_pptx}
\alias{modelo_pptx}
\title{Abrir modelo de apresentacao do pacote}
\usage{
modelo_pptx()
}
\value{
Objeto \code{pptx} (classe do pacote officer).
}
\description{
Esta funcao carrega o arquivo .pptx incluído no pacote
e retorna um objeto pptx pronto para ser manipulado com o pacote officer.
}
\examples{
ppt <- modelo_pptx()
print(ppt)
}
