% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monta_ppt_normal.R
\name{monta_ppt_normal}
\alias{monta_ppt_normal}
\title{Monta um slide PowerPoint simples contendo uma tabela e opcionalmente um mapa e um grafico}
\usage{
monta_ppt_normal(
  ft_tests,
  tem_test,
  title_tab = "Tabela",
  titulo = "Titulo slide",
  caminho_arquivo = "saida_apresentacao.pptx",
  caminho_grafico = NULL,
  caminho_map = NULL
)
}
\arguments{
\item{ft_tests}{Flextable a ser inserido no slide.}

\item{tem_test}{Lógico indicando se tabela deve ser exibida.}

\item{title_tab}{Titulo da tabela}

\item{titulo}{Titulo do slide.}

\item{caminho_arquivo}{Local de salvamento do arquivo \code{.pptx}.}

\item{caminho_grafico}{Caminho opcional para grafico.}

\item{caminho_map}{Caminho opcional para mapa.}
}
\value{
Arquivo \code{.pptx} salvo no caminho informado.
}
\description{
Monta um slide PowerPoint simples contendo uma tabela e opcionalmente um mapa e um grafico
}
