% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monta_ppt_process.R
\name{monta_ppt_process}
\alias{monta_ppt_process}
\title{Monta um slide PowerPoint simples contendo uma tabela e opcionalmente um mapa}
\usage{
monta_ppt_process(
  ft,
  tem_tabela,
  process = NULL,
  titulo = "Tabela Gerada no R",
  caminho_arquivo = "saida_apresentacao.pptx",
  caminho_map = NULL
)
}
\arguments{
\item{ft}{Flextable a ser inserido no slide.}

\item{tem_tabela}{Lógico indicando se tabela deve ser exibida.}

\item{process}{processo do qual sera gerado o slide, pode ser "winsorize" ou "BoxCox"}

\item{titulo}{Título do slide.}

\item{caminho_arquivo}{Local de salvamento do arquivo \code{.pptx}.}

\item{caminho_map}{Caminho opcional para mapa.}
}
\value{
Arquivo \code{.pptx} salvo no caminho informado.
}
\description{
Monta um slide PowerPoint simples contendo uma tabela e opcionalmente um mapa
}
