% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{mun_shp}
\alias{mun_shp}
\title{Shape de municípios}
\format{
Um objeto \code{sf} com X linhas e Y colunas.
}
\source{
IBGE
}
\usage{
mun_shp
}
\description{
Objeto \code{sf} com limites municipais.
}
\keyword{datasets}
