% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{plotAlphaCronbach}
\alias{plotAlphaCronbach}
\title{Figura do Alpha de Cronbach}
\usage{
plotAlphaCronbach(alpha_df, alpha_total, nfile = "output.png")
}
\arguments{
\item{alpha_df}{Data frame com colunas \code{Indicador} e \code{Alpha_if_Dropped} (como retornado em \code{ADPAlphaCron}).}

\item{alpha_total}{Valor total do alfa de Cronbach.}

\item{nfile}{Nome do arquivo de saída (PNG).}
}
\value{
Nenhum objeto retornado. Salva a figura em arquivo PNG.
}
\description{
Gera gráfico mostrando o impacto da remoção de cada variável sobre o alfa de Cronbach.
Variáveis invertidas são indicadas com um asterisco (*).
}
\examples{
res <- correl_ind(mtcars)
plotAlphaCronbach(res$AlphaCrobach$alpha_df, res$AlphaCrobach$alpha_total, "Alpha.png")

}
