% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPBoxCox.R
\name{sfunc_bxcx}
\alias{sfunc_bxcx}
\title{Aplicar transformada Box-Cox (ou variacoes) a um unico indicador}
\usage{
sfunc_bxcx(Y, metodo)
}
\arguments{
\item{Y}{Vetor numerico contendo os valores do indicador ou indice.}

\item{metodo}{Definicao do metodo de transformacao a ser utilizado:

\itemize{
\item \code{"forecast"}: Usa \code{forecast::BoxCox} com lambda estimado por \code{forecast::BoxCox.lambda}.
\item \code{"COINr"}: Usa \code{COINr::boxcox} com lambda estimado por \code{forecast::BoxCox.lambda}.
\item \code{"yeojohnson"}: Usa \code{bestNormalize::yeojohnson}.
}}
}
\value{
Vetor numerico com os valores transformados, mantendo a localizacao dos \code{NA} originais.
}
\description{
Aplica a transformada Box-Cox, Yeo-Johnson ou metodo equivalente a um unico vetor numerico.

Com suporte para diferentes metodos de transformacao.

Remove valores ausentes antes do calculo.

Ajustando valores iguais a zero
para evitar problemas de logaritmo.
}
\details{
\itemize{
\item Valores \code{NA} sao ignorados na aplicacao da transformada e preservados na saida.
\item Nos valores iguais a zero sao adicionado +0.001 para evitar erros numericos.
}
}
\examples{
#Vetor de exemplo
dados <- c(25, 30, 28, 35, 40, 28, NA, 40,25,14,NA,33,56,23,NA,27,20, 35,71)
# Aplicando Box-Cox com pacote forecast
bxcx.fcs <- sfunc_bxcx(dados, "forecast")

# Aplicando Box-Cox com o COINr
bxcx.coin <- sfunc_bxcx(dados, "COINr")

# Aplicando Yeo-Johnson
bxcx.yjn <- sfunc_bxcx(dados, "yeojohnson")

par(mfrow=c(1,3)) ; par(mar=c(5,5,6,1))

plot(index,bxcx.fcs,
     main = "Transformacao BoxCox \n via pacote Forecast",
     cex.axis=1.5,las=1,ylab="Transformada BoxCox",
     xlab="Reference", cex.lab=1.8,cex.main=2) 

plot(index,bxcx.coin,
     main = "Transformacao BoxCox \n via pacote COINr",
     cex.axis=1.5,las=1,ylab="Transformada BoxCox",
     xlab="Reference",cex.lab=1.8,cex.main=2) 

plot(index,bxcx.yjn,
     main = "Transformacao BoxCox \n via pacote bestNormalize::Yeojohnson",
     cex.axis=1.5,las=1,ylab="Transformada BoxCox",
     xlab="Reference",cex.lab=1.8,cex.main=2)
}
