% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_normal.R
\name{slides_normal}
\alias{slides_normal}
\title{Gera um slide PowerPoint contendo uma tabela genérica e opcionalmente um mapa e um gtrafico}
\usage{
slides_normal(
  datain = datain,
  titulo = "Titulo Slide",
  title_tab = "Titulo Tabela",
  caminho_arquivo = "saida_apresentacao.pptx",
  caminho_map = NULL,
  caminho_grafico = NULL
)
}
\arguments{
\item{datain}{Data frame a ser aplicado os textes.}

\item{titulo}{Titulo do slide.}

\item{title_tab}{Titulo da tabela}

\item{caminho_arquivo}{Caminho de saida do arquivo \code{.pptx}.}

\item{caminho_map}{Caminho opcional para imagem de mapa.}

\item{caminho_grafico}{Caminho opcional para o grafico.}
}
\value{
Um arquivo PowerPoint salvo no caminho especificado.
}
\description{
Função de alto nível que cria o flextable e monta o slide em um único passo.
}
\examples{
\dontrun{
slides_normal(datain=result$Data_Bxc$data[[icode]],
 titulo          = "Titulo Slide",
 title_tab       = "Testes Estatísticos",
 caminho_arquivo = "saida_apresentacao.pptx"),
 caminho_map     = "mapa_normalizado_T1.png" ,
 caminho_grafico  ="grafico_final_T1.png" )
}

}
