% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_winsorize.R
\name{slides_winsorize}
\alias{slides_winsorize}
\title{Gera um slide PowerPoint contendo uma tabela genérica e opcionalmente um mapa}
\usage{
slides_winsorize(
  tabela1,
  titulo = "Tabela Gerada no R",
  caminho_arquivo = "saida_apresentacao.pptx",
  caminho_map = NULL
)
}
\arguments{
\item{tabela1}{Data frame a ser exibido.}

\item{titulo}{Título do slide.}

\item{caminho_arquivo}{Caminho de saída do arquivo \code{.pptx}.}

\item{caminho_map}{Caminho opcional para imagem de mapa.}
}
\value{
Um arquivo PowerPoint salvo no caminho especificado.
}
\description{
Função de alto nível que cria o flextable e monta o slide em um único passo.
}
\examples{
\dontrun{
slides_generico(mtcars)
}

}
