% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{total.na}
\alias{total.na}
\title{Contagem de Valores Ausentes}
\usage{
total.na(x)
}
\arguments{
\item{x}{Vetor, matriz ou coluna de dataframe a ser avaliada.}
}
\value{
Um número inteiro indicando a quantidade de elementos ausentes (\code{NA}).
}
\description{
Conta a quantidade de valores \code{NA} em um vetor numérico ou dataframe.
}
\examples{
total.na(c(1, NA, 3, NA, 5))
# Retorna: 2

}
