% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{uf_shp}
\alias{uf_shp}
\title{Shape de unidades federativas}
\format{
Objeto \code{sf}.
}
\source{
IBGE
}
\usage{
uf_shp
}
\description{
Objeto sf com unidades da federação.
}
\examples{
head(uf_shp)
}
\keyword{datasets}
