% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winzorise.r
\name{wins_par}
\alias{wins_par}
\title{Gerar resumo sobre os limites do Winsorize de unico indicador.
,}
\usage{
wins_par(indice, classe, nome)
}
\arguments{
\item{indice}{Um vetor contendo os dados .}

\item{classe}{classe do indicador (ex: "Numerico", "Cluster", "Score")."}

\item{nome}{Vetor com os nomes descritivos de cada coluna de \code{data}.}
}
\description{
Esta funcao gera os limites inferiores e superiores para a aplicacao do winzorise num indicador especifico,
com suporte para indicadores do tipo cluster.
,
}
