% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winzorise.r
\name{winsorize_apply}
\alias{winsorize_apply}
\title{Aplicar Winsorization a um conjunto de dados}
\usage{
winsorize_apply(dataset = NULL, metadados = NULL)
}
\arguments{
\item{dataset}{\code{data.frame} contendo os dados originais.}

\item{metadados}{\code{data.frame} de metadados, conforme exigido por \code{winsorize_data}.}
}
\value{
Lista com dois elementos:
\describe{
\item{Resumo}{\code{data.frame} com os limites e informacoes do Winsorization.}
\item{dataset}{\code{data.frame} com os dados ajustados.}
}
}
\description{
Aplica a funcao \code{winsorize_data} para calcular limites de Winsorization
e depois substitui os valores dos indicadores de acordo com esses limites,
considerando variaveis numericas.
}
\details{
\itemize{
\item A funcao espera que a saida de \code{winsorize_data} contenha:
\itemize{
\item Coluna 1: nome ou indice da variavel.
\item Coluna 2: tipo ("Numerico", "Descricao", "Score", "Grupo 1", "Grupo 2").
\item Colunas \code{LSUP} e \code{LINF}: limites superiores e inferiores.
}
}
}
\examples{
# Exemplo ficticio (assumindo que winsorize_data e winsorize_info ja existam)
 #dados <- data.frame(
  # var1 = c(1, 2, 3, 100),
  #var2 = c(10, 20, 30, 40)
  #)
data("metadadosN7", package = "reprodutibilidade")

data("datasetN7", package = "reprodutibilidade")

data_winsor <- winsorize_apply(dataset=datasetN7,
                             metadados=metadadosN7)

}
