% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winzorise.r
\name{winsorize_data}
\alias{winsorize_data}
\title{Aplicar Winsorization a variaveis numericas}
\usage{
winsorize_data(dataset = NULL, metadados = NULL)
}
\arguments{
\item{dataset}{\code{data.frame} com os dados originais.}

\item{metadados}{\code{data.frame} com metadados das variaveis. Deve conter colunas:
\itemize{
\item \code{CLASSE} - Tipo da variavel ("Numerico").
\item \code{CODE} - Nome da coluna correspondente em \code{dataset}.
}}
}
\value{
Um \code{data.frame} com o resumo da Winsorization de cada variavel.
}
\description{
Aplica o processo de Winsorization a colunas numericas
}
\details{
Esta funcao utiliza a funcao auxiliar \code{winsorize_info} para calcular os limites de Winsorization
e retorna um resumo para cada variavel processada.
\itemize{
\item As variaveis sao retornadas na ordem definida por \code{metadados$CODE}.
}
}
\examples{
# Exemplo ficticio (assumindo que winsorize_info ja esteja implementada)
dataset <- data.frame(
  var1 = c(1, 2, 3, 100),
  var2 = c(10, 20, 30, 40)
)

metadados <- data.frame(
  CLASSE = c("Numerico", "Numerico"),
  CODE = c("var1", "var2")
)


winsorize_data(dataset, metadados)

}
