% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winzorise.r
\name{winsorize_info}
\alias{winsorize_info}
\title{Gerar resumo sobre os limites do Winsorize de unico indicador.
,}
\usage{
winsorize_info(indice, classe_label = NULL, var_name = NULL)
}
\arguments{
\item{indice}{Um vetor contendo os dados de entrada .}

\item{classe_label}{classe do indice (ex: "Numerico", "Cluster", "Score")."}

\item{var_name}{nome da variavel.}
}
\description{
Esta funcao gera os limites inferiores e superiores para a aplicacao do winzorise num indicador especifico,
com suporte para indicadores do tipo cluster.
,
}
