% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{ADPcorrel}
\alias{ADPcorrel}
\title{Correlação de Spearman entre Variáveis}
\usage{
ADPcorrel(X)
}
\arguments{
\item{X}{Data frame ou matriz numérica com variáveis contínuas.}
}
\value{
Uma lista contendo:
\itemize{
\item \code{cor_mat}: matriz de correlação absoluta.
\item \code{p_mat}: matriz de valores-p da correlação.
\item \code{cor_summary}: resumo indicando a correlação máxima de cada variável
e o respectivo par.
}
}
\description{
Calcula a matriz de correlação de Spearman e identifica, para cada variável,
qual é a outra variável mais correlacionada.
}
\examples{
res <- ADPcorrel(mtcars)
head(res$cor_summary)

}
