% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPNormalise.R
\name{ADPNormalise}
\alias{ADPNormalise}
\title{Normalizar dados para o intervalo \code{[0 , 1]}}
\usage{
ADPNormalise(iData)
}
\arguments{
\item{iData}{\code{data.frame} ou \code{matrix} contendo os dados.
Valores nao numericos serao ignorados e retornados sem modificacao.}
}
\value{
Lista com:
\describe{
\item{iData}{\code{data.frame} com colunas normalizadas (numericas) ou originais (nao numericas).}
}
}
\description{
Aplica a normalizacao min-max para cada coluna numerica de um data.frame ou matriz.
}
\details{
A normalizacao min-max e dada por:
\deqn{ X' = (X - \min(X)) / (\max(X) - \min(X)) }

Colunas constantes (mesmo valor para todos) sao retornadas como 0.
}
\examples{
df <- data.frame(a = 1:5, b = c(2,2,2,2,2), c = letters[1:5])
ADPNormalise(df)

}
