% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPresumo.R
\name{criar_resumo}
\alias{criar_resumo}
\title{Gerar resumo estatístico para um indicador ou índice específico,
,}
\usage{
criar_resumo(data, class_type, cluster = NULL, name)
}
\arguments{
\item{data}{Um vetor contendo indicador.}

\item{class_types}{classe do indicador que será avaliado (ex: "Numérico", "Cluster", "Score")."}

\item{clusters}{Um vetor com os grupos de cluster (obrigatório se usar tipo "Cluster")."}

\item{names}{o nome do indicador.}
}
\value{
Um tibble (tabela moderna do R, semelhante a um data.frame) com os resumos estatísticos completos para cada variável.
\describe{
\item{Min}{Valor mínimo}
\item{Quartil 1}{Primeiro Quartil}
\item{Mediana}{Valor mediano}
\item{Quartil 3}{Terceiro Quartil}
\item{Max}{Valor máximo}
\item{Outliers_Per}{Percentual de outliers do indicador/indice}
\item{NAs}{Total de valores ausentes (NAs)}
\item{Valores_Unicos}{Total de valores únicos}}
}
\description{
Esta função aplica resumos estatísticos a múltiplas variáveis de um data frame,
com suporte para indicadores do tipo cluster.
,
}
\examples{
# Criando um data frame de exemplo
dados <- data.frame(Pessoas = c(25, 30, 28, 35, 40, 28, 35, 40),
                   Cluster = c("1", "2", "1", "1", "2","2","2","1"))

# Chamando a função com parâmetros fictícios
criar_resumo(dados$Pessoas,
            "Numérico",
             dados$cluster,
             "Pessoas")

}
