% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPCorrel.R
\name{FigCorrelPlot}
\alias{FigCorrelPlot}
\title{Mapa de Correlacao}
\usage{
FigCorrelPlot(Y, tipo = "Total", nfile = "output.png", visivel = TRUE)
}
\arguments{
\item{Y}{Matriz de correlacao.}

\item{tipo}{Tipo de correlacao a ser mostrada: "Total" (parte inferior) ou "Parcial" (parte superior).}

\item{nfile}{Nome do arquivo de saida (PNG).}

\item{visivel}{logical, se TRUE o grafico e exibido no prompt.}
}
\value{
Nenhum objeto retornado. Salva a figura em arquivo PNG.
}
\description{
Gera figura de correlacao (com \code{corrplot}) para todas as variaveis.
}
\examples{
res <- correl_ind(mtcars)
FigCorrelPlot(res$Correl, tipo = "Total", nfile = "Correl.png",visivel=TRUE)

}
