% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPGraficos.R
\name{grafico_final}
\alias{grafico_final}
\title{Gerar Grafico Combinado (Dispersao + Histograma)}
\usage{
grafico_final(
  dtset_proc,
  dtset_norm,
  nome_arquivo = "grafico_combinado.png",
  largura = 10,
  altura = 5,
  dpi = 25,
  icode = icode,
  fsize = 16,
  plot = FALSE
)
}
\arguments{
\item{dtset_proc}{Data frame contendo os dados nao normalizados,
usadas para o grafico de dispersao.}

\item{dtset_norm}{Data frame contendo os dados normalizados a ser usada no histograma e no grafico de dispersao
(coluna indicada em \code{icode}).}

\item{nome_arquivo}{Nome do arquivo de saida (\code{.png}). Default: "grafico_combinado.png".}

\item{largura}{Largura da imagem em polegadas. Default: 10.}

\item{altura}{Altura da imagem em polegadas. Default: 5.}

\item{dpi}{Resolucao (pontos por polegada). Default: 25.}

\item{icode}{String com o nome da variavel em \code{df} que sera utilizada no histograma.
Default: \code{icode}.}

\item{fsize}{Tamanho da fonte para titulos, legendas e eixos. Default: 16.}

\item{plot}{Logical; se TRUE, plota na tela. Se FALSE, apenas retorna o objeto.}
}
\value{
Nenhum objeto e retornado. Um arquivo PNG contendo o grafico combinado e salvo no diretorio de trabalho.
}
\description{
Esta funcao cria um grafico composto por:
(i) um grafico de dispersao entre valores normalizados e brutos; e
(ii) um histograma dos dados normalizados com curva de distribuicao normal ajustada.
Os dois graficos sao apresentados lado a lado em um unico layout e salvos em arquivo PNG.
}
\details{
\itemize{
\item O grafico de dispersao usa os eixos:
\code{x = Normalizado}, \code{y = N_Normalizado}.
\item O histograma inclui uma linha da distribuicao normal ajustada (media e desvio-padrao dos dados).
\item Em caso de erro na geracao do histograma (ex.: coluna nao encontrada), e exibido
um grafico de mensagem de erro no lugar.
}
}
\examples{
\dontrun{
# Exemplo com dados simulados

grafico_final(dtset_proc = result$Data_Bxc$data,
          dtset_norm = result$Data_Normal$dataset, 
            nome_arquivo = "grafico_final_T1.png", 
            largura = 20, 
            altura = 10, 
            dpi = 100,
            icode="MMPD",
            fsize=32)
}

}
\seealso{
\code{\link[ggplot2]{ggplot}},
\code{\link[gridExtra]{arrangeGrob}},
\code{\link[ggplot2]{stat_function}}
}
