% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{mun_shp}
\alias{mun_shp}
\title{Shape de municípios}
\format{
Objeto \code{sf} com colunas de atributos.
}
\source{
IBGE
}
\usage{
mun_shp
}
\description{
Objeto sf com municípios brasileiros, incluindo a coluna \code{CD_MUN}.
}
\examples{
head(mun_shp)
}
\keyword{datasets}
