% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPNormalise.R
\name{sfunc_norm}
\alias{sfunc_norm}
\title{Normalizacao Min-Max}
\usage{
sfunc_norm(Y)
}
\arguments{
\item{Y}{Vetor numerico a ser normalizado.}
}
\value{
Vetor numerico normalizado no intervalo [0, 1], preservando as posicoes originais de \code{NA}.
}
\description{
Normaliza um vetor numerico usando o metodo Min-Max,
transformando seus valores para o intervalo [0, 1].
}
\details{
\itemize{
\item Valores \code{NA} sao ignorados nos calculos de minimo e maximo, mas preservados na saida.
\item Formula utilizada:
\deqn{Y_{norm} = \frac{Y - \min(Y)}{\max(Y) - \min(Y)}}
}
}
\examples{
# Vetor de exemplo
dados <- c(10, 20, 30, NA, 40)

# Normalizacao Min-Max
sfunc_norm(dados)

}
