% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{uf_shp}
\alias{uf_shp}
\title{Shape de unidades federativas}
\format{
Um objeto \code{sf} com X linhas e Y colunas.
}
\source{
IBGE
}
\usage{
uf_shp
}
\description{
Objeto \code{sf} com limites das UFs.
}
\keyword{datasets}
